Moved MockLambdaContext to infra/__init__.py
